[ADD] Added mapping sensitive predictor.
