+++ src/parameters.hpp
@@ -126,6 +126,7 @@ typedef enum{
 // strategy is "LEARNING".
 typedef enum{
     STRATEGY_PREDICTION_REGRESSION_LINEAR = 0,
+    STRATEGY_PREDICTION_REGRESSION_LINEAR_MAPPING,
     STRATEGY_PREDICTION_MISHRA
 }StrategyPrediction;
