+++ src/predictors.cpp
@@ -493,6 +493,62 @@ double PredictorLinearRegression::predict(const KnobsValues& values, double band
 
     return res;
 }
+
+
+/*************************************************/
+/*        PredictorLinearRegressionMapping       */
+/*************************************************/
+PredictorLinearRegressionMapping::PredictorLinearRegressionMapping(PredictorType type,
+                          const Parameters& p,
+                          const Configuration& configuration,
+                          const Smoother<MonitoredSample>* samples):
+            Predictor(type, p, configuration, samples){
+    for(size_t i = 0; i < MAPPING_TYPE_NUM; i++){
+        _predictors[i] = new PredictorLinearRegression(type, p, configuration, samples);
+    }
+}
+
+PredictorLinearRegressionMapping::~PredictorLinearRegressionMapping(){
+    for(size_t i = 0; i < MAPPING_TYPE_NUM; i++){
+        delete _predictors[i];
+    }
+}
+
+void PredictorLinearRegressionMapping::clear(){
+    for(size_t i = 0; i < MAPPING_TYPE_NUM; i++){
+        _predictors[i]->clear();
+    }
+}
+
+bool PredictorLinearRegressionMapping::readyForPredictions(){
+    for(size_t i = 0; i < MAPPING_TYPE_NUM; i++){
+        if(!_predictors[i]->readyForPredictions()){
+            return false;
+        }
+    }
+    return true;
+}
+
+void PredictorLinearRegressionMapping::refine(){
+    _predictors[(MappingType) _configuration.getRealValue(KNOB_TYPE_MAPPING)]->refine();
+}
+
+void PredictorLinearRegressionMapping::prepareForPredictions(){
+    for(size_t i = 0; i < MAPPING_TYPE_NUM; i++){
+        _predictors[i]->prepareForPredictions();
+    }
+}
+
+double PredictorLinearRegressionMapping::predict(const KnobsValues& configuration,
+                                                 double bandwidthIn){
+    double realValue = 0.0;
+    if(configuration.areRelative()){
+        _configuration.getKnob(KNOB_TYPE_MAPPING)->getRealFromRelative(configuration[KNOB_TYPE_MAPPING], realValue);
+    }else{
+        realValue = configuration[KNOB_TYPE_MAPPING];
+    }
+    return _predictors[(MappingType) realValue]->predict(configuration, bandwidthIn);
+}
     
 /**************** PredictorSimple ****************/
